rename States and Edges classes to SuperstructureStates and SuperstructureEdges and move EdgeCommands to the Edges class
