+++ src/main/java/frc/robot/subsystems/v2_Redundancy/superstructure/V2_RedundancySuperstructureEdges.java
@@ -1,11 +1,24 @@
 package frc.robot.subsystems.v2_Redundancy.superstructure;
 
-import frc.robot.subsystems.v2_Redundancy.superstructure.V2_RedundancyStates.SuperstructureStates;
+import edu.wpi.first.wpilibj2.command.Command;
+import edu.wpi.first.wpilibj2.command.Commands;
+import frc.robot.subsystems.v2_Redundancy.superstructure.V2_RedundancySuperstructureStates.SuperstructureStates;
+import frc.robot.subsystems.v2_Redundancy.superstructure.elevator.V2_RedundancyElevator;
+import frc.robot.subsystems.v2_Redundancy.superstructure.funnel.V2_RedundancyFunnel;
+import frc.robot.subsystems.v2_Redundancy.superstructure.intake.V2_RedundancyIntake;
+import frc.robot.subsystems.v2_Redundancy.superstructure.intake.V2_RedundancyIntakeConstants.IntakeExtensionState;
+import frc.robot.subsystems.v2_Redundancy.superstructure.intake.V2_RedundancyIntakeConstants.IntakeRollerState;
+import frc.robot.subsystems.v2_Redundancy.superstructure.manipulator.V2_RedundancyManipulator;
+import frc.robot.subsystems.v2_Redundancy.superstructure.manipulator.V2_RedundancyManipulatorConstants.ArmState;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import lombok.Builder;
+import lombok.Getter;
+import org.jgrapht.Graph;
+import org.jgrapht.graph.DefaultEdge;
 
-public class V2_RedundancyEdges {
+public class V2_RedundancySuperstructureEdges {
 
   public static final ArrayList<Edge> NoneEdges = new ArrayList<>();
   public static final ArrayList<Edge> AlgaeEdges = new ArrayList<>();
@@ -23,7 +36,75 @@ public class V2_RedundancyEdges {
     }
   }
 
-  public static void createEdges() {
+  public enum AlgaeEdge {
+    NONE,
+    NO_ALGAE,
+    ALGAE
+  }
+
+  @Builder(toBuilder = true)
+  @Getter
+  public static class EdgeCommand extends DefaultEdge {
+    private final Command command;
+    @Builder.Default private final AlgaeEdge algaeEdgeType = AlgaeEdge.NONE;
+  }
+
+  private static Command getEdgeCommand(SuperstructureStates from, SuperstructureStates to, 
+      V2_RedundancyElevator elevator,
+      V2_RedundancyManipulator manipulator,
+      V2_RedundancyFunnel funnel,
+      V2_RedundancyIntake intake) {
+    V2_RedundancySuperstructurePose pose = to.getPose();
+
+    if (from == SuperstructureStates.INTAKE_FLOOR) {
+      return Commands.parallel(
+          pose.asCommand(elevator, manipulator, funnel, intake),
+          Commands.run(() -> intake.setRollerGoal(IntakeRollerState.OUTTAKE))
+              .withTimeout(1)
+              .andThen(Commands.runOnce(() -> intake.setRollerGoal(IntakeRollerState.STOP))));
+    }
+
+    if ((to == SuperstructureStates.INTAKE_REEF_L2 || to == SuperstructureStates.INTAKE_REEF_L3)
+        && (from != SuperstructureStates.STOW_UP || from != SuperstructureStates.BARGE)) {
+      return Commands.sequence(
+          pose.setElevatorHeight(elevator)
+              .alongWith(
+                  Commands.runOnce(() -> manipulator.setAlgaeArmGoal(ArmState.STOW_DOWN))
+                      .alongWith(manipulator.waitUntilAlgaeArmAtGoal()))
+              .alongWith(pose.setFunnelState(funnel).alongWith(pose.setIntakeState(intake))),
+          pose.setArmState(manipulator));
+    }
+
+    if (to == SuperstructureStates.FLOOR_ACQUISITION) {
+      return Commands.deadline(
+          pose.asCommand(elevator, manipulator, funnel, intake),
+          Commands.runOnce(() -> intake.setRollerGoal(IntakeRollerState.INTAKE)));
+    }
+
+    if (to == SuperstructureStates.INTERMEDIATE_WAIT_FOR_ARM
+        || (from == SuperstructureStates.FLOOR_ACQUISITION && to == SuperstructureStates.STOW_DOWN)
+        || to == SuperstructureStates.STOW_UP) {
+      return pose.setArmState(manipulator)
+          .andThen(
+              pose.setIntakeState(intake)
+                  .alongWith(pose.setElevatorHeight(elevator))
+                  .alongWith(pose.setFunnelState(funnel)));
+    }
+    if (to == SuperstructureStates.INTERMEDIATE_WAIT_FOR_ELEVATOR) {
+      return pose.setElevatorHeight(elevator)
+          .andThen(
+              pose.setIntakeState(intake)
+                  .alongWith(pose.setArmState(manipulator), pose.setFunnelState(funnel)));
+    }
+    if (to == SuperstructureStates.L1 && from == SuperstructureStates.SCORE_L1) {
+      return pose.asCommand(elevator, manipulator, funnel, intake)
+          .andThen(Commands.runOnce(() -> intake.setExtensionGoal(IntakeExtensionState.L1_EXT)));
+    }
+
+    return pose.asCommand(elevator, manipulator, funnel, intake); // does all subsystem poses in paralell
+  }
+
+  private static void createEdges() {
 
     // CORAL-RELATED STATES
     List<SuperstructureStates> coralLevels =
@@ -249,5 +330,32 @@ public class V2_RedundancyEdges {
         new Edge(SuperstructureStates.STOW_UP, SuperstructureStates.FUNNEL_CLOSE_WITH_STOW_UP));
     AlgaeEdges.add(
         new Edge(SuperstructureStates.FUNNEL_CLOSE_WITH_STOW_UP, SuperstructureStates.STOW_UP));
+  }
+
+  private static void addEdges(
+      Graph<SuperstructureStates, EdgeCommand> graph, List<Edge> edges, AlgaeEdge type, V2_RedundancyElevator elevator,
+      V2_RedundancyManipulator manipulator,
+      V2_RedundancyFunnel funnel,
+      V2_RedundancyIntake intake) {
+    for (Edge edge : edges) {
+      graph.addEdge(
+          edge.from(),
+          edge.to(),
+          EdgeCommand.builder()
+              .command(getEdgeCommand(edge.from(), edge.to(), elevator, manipulator, funnel, intake))
+              .algaeEdgeType(type)
+              .build());
+    }
+  }
+
+  public static void addEdges(Graph<SuperstructureStates, EdgeCommand> graph, 
+      V2_RedundancyElevator elevator,
+      V2_RedundancyManipulator manipulator,
+      V2_RedundancyFunnel funnel,
+      V2_RedundancyIntake intake) {
+    createEdges();
+    addEdges(graph, NoneEdges, AlgaeEdge.NONE, elevator, manipulator, funnel, intake);
+    addEdges(graph, NoAlgaeEdges, AlgaeEdge.NO_ALGAE, elevator, manipulator, funnel, intake);
+    addEdges(graph, AlgaeEdges, AlgaeEdge.ALGAE, elevator, manipulator, funnel, intake);
   }
 }
