+++ src/main/java/frc/robot/subsystems/v2_Redundancy/superstructure/V2_RedundancySuperstructureStates.java
@@ -13,7 +13,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
-public class V2_RedundancyStates {
+public class V2_RedundancySuperstructureStates {
   public static final List<SuperstructureStates> Actions =
       List.of(
           SuperstructureStates.INTAKE,
